* added some output to multi_fit.sh
git-svn-id: file:///Users/demichel/SVNREPOS/CODE@671 fea3b138-1bf5-0310-9da5-cbb00e418fd0

+++ SCRIPTS/multi_fit.sh
@@ -49,42 +49,52 @@ cd ..
 continue
 fi
 cd Phi$2
+echo "Processing " $X0 " Phi" $2 
 #echo $RCMD " " `pwd`
 if [ -e Cn.dat ]
 then
-RCMD=`echo "$PERC/fitSE.C(\"Cn.dat\",2,$CNBEG,1000,1)"`
+echo -n "Fitting Cn.dat..."
+RCMD=`echo "$PERC/fitSE.C(\"Cn.dat\",2,$CNBEG,1000,0)"`
 $ROOTEXE -b -q $RCMD > $LF
 TAU=`cat $LF | tail -4 | awk '{if ($2=="p1") print $3}'`
 BETA=`cat $LF | tail -4 | awk '{if ($2=="p2") print $3}'`
 CHISQ=`cat $LF | tail -4 | awk '{if ($1=="CHISQUARE:") print $2}'`
 echo $X0 $TAU $CHISQ >> ../../$CNTAUVSX0
 echo $X0 $BETA $CHISQ >> ../../$CNBETAVSX0
+echo "done"
 fi
 #
-FQS=`ls -r -1 Fqs*max | tail -1` 
+FQS=`ls -r -1 Fqs*max | tail -1 2> /dev/null` 
 if [ "$FQS" != "" ]
 then
-RCMD=`echo "$PERC/fitSE.C(\"$FQS\",2,$FQSBEG,1000,1)"`
+echo -n "Fitting " $FQS "..."
+RCMD=`echo "$PERC/fitSE.C(\"$FQS\",2,$FQSBEG,1000,0)"`
 $ROOTEXE -b -q $RCMD > $LF
 TAU=`cat $LF | tail -4 | awk '{if ($2=="p1") print $3}'`
 BETA=`cat $LF | tail -4 | awk '{if ($2=="p2") print $3}'`
 CHISQ=`cat $LF | tail -4 | awk '{if ($1=="CHISQUARE:") print $2}'`
 echo $X0 $TAU $CHISQ >> ../../$FQSTAUVSX0
 echo $X0 $BETA $CHISQ >> ../../$FQSBETAVSX0
+echo "done"
 fi
 #
 FQC=`ls -r -1 N-sqt*max | tail -1` 
 if [ "$FQC" != "" ]
 then
-RCMD=`echo "$PERC/fitSE.C(\"$FQC\",2,$FQCBEG,1000,1)"`
+echo -n "Fitting " $FQC "..."
+RCMD=`echo "$PERC/fitSE.C(\"$FQC\",2,$FQCBEG,1000,0)"`
 $ROOTEXE -b -q $RCMD > $LF
 TAU=`cat $LF | awk '{if ($2=="p1") print $3}'`
 BETA=`cat $LF | awk '{if ($2=="p2") print $3}'`
 CHISQ=`cat $LF | awk '{if ($1=="CHISQUARE:") print $2}'`
 echo $X0 $TAU $CHISQ >> ../../$FQCTAUVSX0
 echo $X0 $BETA $CHISQ >> ../../$FQCBETAVSX0
+echo "done"
 fi
+if [ -e $LF ] 
+then
 rm $LF
+fi
 cd ..
 cd ..
 done
